refactor(scene-detail): formalize scene detail module boundary and internal layers
