+++ src/app/routes/AppRoutes.tsx
@@ -1,11 +1,11 @@
 import { Navigate, Route, Routes } from 'react-router-dom'
 import { Layout } from '@components/Layout'
+import { SceneDetailPage } from '@modules/scene-detail'
 import { CreateScenePage } from '@pages/CreateScenePage'
 import { HomePage } from '@pages/HomePage'
 import { LoginPage } from '@pages/LoginPage'
 import { MyScenesPage } from '@pages/MyScenesPage'
 import { RegisterPage } from '@pages/RegisterPage'
-import { SceneDetailPage } from '@pages/SceneDetailPage'
 import { ScenesPage } from '@pages/ScenesPage'
 import { SettingsPage } from '@pages/SettingsPage'
 import { GuestOnlyRoute, ProtectedRoute } from './RouteGuards'
