+++ src/modules/scene-detail/index.ts
@@ -1,0 +1,1 @@
+export { SceneDetailPage } from './SceneDetailPage'
