+++ src/modules/scene-detail/ui/PlaylistPanel.tsx
