+++ src/modules/scene-detail/ui/SceneDetailState.tsx
