+++ src/modules/scene-detail/ui/VoteButton.tsx
