+++ src/modules/scene-detail/ui/playlistPanel.css
