+++ src/modules/scene-detail/ui/playlistPanelIcons.tsx
