+++ src/pages/index.ts
@@ -4,6 +4,6 @@ export { LoginPage } from './LoginPage'
 export { MyScenesPage } from './MyScenesPage'
 export { PlaceholderPage } from './PlaceholderPage'
 export { RegisterPage } from './RegisterPage'
-export { SceneDetailPage } from './SceneDetailPage'
+export { SceneDetailPage } from '@modules/scene-detail'
 export { ScenesPage } from './ScenesPage'
 export { SettingsPage } from './SettingsPage'
